# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Water Pump Dashboard API
  version: 1.0.0
  description: API for monitoring and controlling water pump systems
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /pumps/{id}/readings:
    get:
      tags:
        - Readings
      summary: Get pump readings
      description: Retrieve readings for a specific pump within a date range
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PumpId'
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PumpReading'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /pumps/{id}/controls:
    post:
      tags:
        - Controls
      summary: Send pump control command
      description: Send a control command to a pump
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PumpId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendPumpCommandInput'
      responses:
        '201':
          description: Command sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PumpControl'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PumpId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Pump ID
  schemas:
...
    PumpReading:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pump_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        pressure:
          type: number
        flow_rate:
          type: number
        temperature:
          type: number
        status:
          type: string
          enum: [running, stopped, error, unknown]
        power_consumption:
          type: number
        created_at:
          type: string
          format: date-time
...
    SendPumpCommandInput:
      type: object
      required:
        - pump_id
        - command_type
      properties:
        pump_id:
          type: string
          format: uuid
        command_type:
          type: string
          enum: [start, stop, set_pressure, set_flow_rate]
        command_value:
          type: string
          nullable: true
    PumpControl:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pump_id:
          type: string
          format: uuid
        command_type:
          type: string
        command_value:
          type: string
          nullable: true
        status:
          type: string
        executed_at:
          type: string
          format: date-time
          nullable: true
        created_by:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: array
          items:
            type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
              retryAfter:
                type: integer